Create Zero Ones Equal One Zeros
Kulyash believes in equality.
Given an integer N, output a binary string of length N such that:

The count of 01 subsequences in the string is equal to the count of 10 subsequences;
The string has at least one occurrence of 0 as well as 1.
If multiple such strings exist, print any. Also, it is guaranteed that corresponding to the given input, an answer always exists.